requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-celery==3.1.17
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-filter==25.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gevent==24.11.1
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
uuid==1.30
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
